pyproject.toml
[tool.poetry]
name = "oz-django"
version = "0.1.0"
description = ""
authors = ["wkdsh21 <wkdsh21@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
pymysql = "^1.1.1"
cryptography = "^43.0.3"
django-ninja = "^1.3.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"
django-stubs = "^5.1.1"
coverage = "^7.6.4"
types-pymysql = "^1.1.0.20241103"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = 3.12
strict = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "oz_django.settings"
